Fixed compilation errors in the Blasius boundary layer example

+++ cantera/trunk/samples/cxx/bvp/blasius.cpp
@@ -1,7 +1,6 @@
 /// @file blasius.cpp
 /// The Blasius boundary layer
 
-#include "cantera/Cantera.h"
 #include "BoundaryValueProblem.h"
 
 
@@ -54,7 +53,7 @@ public:
 
     // specify guesses for the initial values. These can be anything
     // that leads to a converged solution.
-    virtual doublereal initialValue(int n, int j) {
+    virtual doublereal initialValue(size_t n, size_t j) {
         switch (n) {
         case 0:
             return 0.1*z(j);
